+++ skoolkit/skoolmacro.py
@@ -462,6 +462,7 @@ def get_macros(writer):
         '#REG': partial(parse_reg, writer.get_reg, writer.case == CASE_LOWER),
         '#SPACE': partial(parse_space, writer),
         '#STR': partial(parse_str, writer),
+        '#TSTATES': partial(parse_tstates, writer),
         '#VERSION': parse_version,
         '#WHILE': partial(parse_while, writer)
     }
@@ -1131,6 +1132,22 @@ def parse_str(writer, text, index, *cwd):
             i, j = re.search(' {2,}', s).span()
             s = s[:i] + f'#SPACE({j-i})' + s[j:]
     return end, s
+
+def parse_tstates(writer, text, index, *cwd):
+    # #TSTATESfirst[,last]
+    end, first, last = parse_ints(text, index, 2, (0,), fields=writer.fields)
+    if last < first:
+        last = first
+    timings = writer.parser.get_instruction_timings(first, last)
+    result = 0
+    for address, timing in timings:
+        if timing is None:
+            raise MacroParsingError(f'Failed to get timing for instruction at {address}')
+        if isinstance(timing, int):
+            result += timing
+        else:
+            result += timing[1]
+    return end, str(result)
 
 def parse_udg(text, index=0, fields=None):
     # #UDGaddr[,attr,scale,step,inc,flip,rotate,mask,tindex,alpha][:addr[,step]][{x,y,width,height}][(fname)]